+++ src/components/HomePage/index.ts
@@ -1,1 +1,1 @@
-export * from './HomePage';
+export * from './Temp';
